python/azure-ai-agents/requirements.txt
opentelemetry-sdk==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
azure-ai-agents==1.1.0b2
azure-ai-projects==1.0.0b11
azure-identity==1.19.0